# readify_frontend/doc/readify.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "readify",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "获取所有工程列表",
        "deprecated": false,
        "description": "",
        "operationId": "getAllProjects",
        "tags": [
          "工程管理"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "example": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJqb2huLmRvZSIsImlkIjo2LCJpYXQiOjE3NDA0MDUyMTUsImV4cCI6MTc0MDQ5MTYxNX0.4Z8XMTuYrNhXOJ7ZuclS0AWSfBoN0Leq9iAMs8NH6-4",
            "schema": {
              "type": "string",
              "default": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJqb2huLmRvZSIsImlkIjo2LCJpYXQiOjE3NDA0MDUyMTUsImV4cCI6MTc0MDQ5MTYxNX0.4Z8XMTuYrNhXOJ7ZuclS0AWSfBoN0Leq9iAMs8NH6-4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListProjectVO"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "创建工程",
        "deprecated": false,
        "description": "",
        "operationId": "createProject",
        "tags": [
          "工程管理"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "example": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJqb2huLmRvZSIsImlkIjo2LCJpYXQiOjE3NDA0MDUyMTUsImV4cCI6MTc0MDQ5MTYxNX0.4Z8XMTuYrNhXOJ7ZuclS0AWSfBoN0Leq9iAMs8NH6-4",
            "schema": {
              "type": "string",
              "default": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJqb2huLmRvZSIsImlkIjo2LCJpYXQiOjE3NDA0MDUyMTUsImV4cCI6MTc0MDQ5MTYxNX0.4Z8XMTuYrNhXOJ7ZuclS0AWSfBoN0Leq9iAMs8NH6-4"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectVO"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProjectVO"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "工程ID",
            "format": "int64",
            "readOnly": true
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "工程名称",
            "examples": [
              "示例工程"
            ]
          },
          "description": {
            "type": "string",
            "description": "工程描述",
            "examples": [
              "这是一个示例工程"
            ]
          },
          "createTime": {
            "type": "integer",
            "description": "创建时间",
            "format": "int64",
            "readOnly": true
          },
          "updateTime": {
            "type": "integer",
            "description": "更新时间",
            "format": "int64",
            "readOnly": true
          }
        },
        "description": "工程视图对象"
      },
...
      "ResultProjectVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "状态码",
            "examples": [
              "200"
            ]
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "examples": [
              "操作成功"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ProjectVO"
          }
        },
        "description": "通用返回结果"
      },
...
      "ResultListProjectVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "状态码",
            "examples": [
              "200"
            ]
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "examples": [
              "操作成功"
            ]
          },
          "data": {
            "type": "array",
            "description": "返回数据",
            "items": {
              "$ref": "#/components/schemas/ProjectVO"
            },
            "nullable": true
          }
        },
        "description": "通用返回结果"
      }
    },
    "securitySchemes": {}
  },
  "servers": []
}